Team datatype fixes

+++ team.go
@@ -6,7 +6,7 @@ type Team struct {
 	CreatedDateUTC       *string  `json:"createdDateUTC,omitempty"`
 	EventID              *int     `json:"eventID,omitempty"`
 	EventName            *string  `json:"eventName,omitempty"`
-	FundraisingGoal      *int     `json:"fundraisingGoal,omitempty"`
+	FundraisingGoal      *float64 `json:"fundraisingGoal,omitempty"`
 	IsInviteOnly         *bool    `json:"isInviteOnly,omitempty"`
 	Links                *Links   `json:"links,omitempty"`
 	Name                 *string  `json:"name,omitempty"`
@@ -14,8 +14,8 @@ type Team struct {
 	NumDonations         *int     `json:"numDonations,omitempty"`
 	NumParticipants      *int     `json:"numParticipants,omitempty"`
 	StreamIsLive         *bool    `json:"streamIsLive,omitempty"`
-	SumDonations         *float32 `json:"sumDonations,omitempty"`
-	SumPledges           *float32 `json:"sumPledges,omitempty"`
+	SumDonations         *float64 `json:"sumDonations,omitempty"`
+	SumPledges           *float64 `json:"sumPledges,omitempty"`
 	TeamID               *int     `json:"teamID,omitempty"`
 	HasActivityTracking  *bool    `json:"hasActivityTracking,omitempty"`
 	SourceTeamID         *int     `json:"sourceTeamID,omitempty"`
